initial pass at reorganizing the lib directory
This is not the final version, but a preview version.

You will see both the old lib/ directory and the new
lib-common/, lib-platforms/, and lib-scripts/ here.
The lib/ directory only remains as something for
you to look at.  The build now operates against
the three new directories.

Another transient piece is the build/split-cordova.sh
file, which is used to actually split the files.  Won't
be in the final commit, but it means I can rebase
this commit and split again safely.

We now also build 'debug' versions of the cordova.js
files.  These can be used in debuggers to provide a
better user experience. See [1] for more info.  Note
that this includes the test runner, running in node,
which provides better diagnostics when bad things
happen.

Changes besides the remix of the directories:

- all new build/packager.js

- Jakefile now calls packager.generate() instead of packager.write()

- test/runner.js changes - it was previously using a special 'build'
  of the lib tree, now it uses a new 'platform' called 'test'.  This
  simplified things a little, and now means the tests are running
  against the same structure as the real libraries, instead of a
  special-purpose-built version.  Also did a few cleanup things
  in runner.js and the associated test/suite.html.

- removed the first test in test.network.js, since it no longer
  passes.  See the comment in the test.

[1] http://pmuellr.blogspot.com/2011/11/debugging-concatenated-javascript-files.html
    scroll down to "cheating with eval() and //@sourceurl"