2nd try at file refactoring
So, this time, I moved the 'lib' directory to 'lib0'
for the time being, and use the script 'build/split-cordova.sh'
to rebuild a new 'lib' directory.

That directory has the following contents:

    lib/android     - platform specific
    lib/blackberry  - platform specific
    lib/common      - common modules
    lib/cordova.js  - base cordova module
    lib/errgen      - platform specific
    lib/ios         - platform specific
    lib/playbook    - platform specific
    lib/scripts     - platform specific
    lib/test        - common scripts (not modules)
    lib/webworks    - platform specific
    lib/wp7         - platform specific

All the files in lib/ with the exception of lib/scripts are
modules, and they're all sitting at the same 'level'.  So, to
build a copy of the source modules for android, you can do
the following:

    cp -R lib/common  gen/android
    cp -R lib/android gen/android

'platform specific' files overwrite any 'common' files,
per the order of the cp commands above.